% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ClassMethod-dcGOR.r
\docType{class}
\name{AdjData-class}
\alias{AdjData}
\alias{AdjData-class}
\title{Definition for VIRTUAL S4 class AdjData}
\value{
Class AdjData
}
\description{
\code{AdjData} is union of other classes: either matrix or dgCMatrix (a
sparse matrix in the package Matrix). It is used as a virtual class
}
\seealso{
\code{\link{Onto-class}}
}
\keyword{S4}
\keyword{classes}

