\docType{methods}
\name{Anno-method}
\alias{Anno-method}
\alias{annoData}
\alias{annoData,Anno-method}
\alias{dData}
\alias{dData,Anno-method}
\alias{dim,Anno-method}
\alias{domainData}
\alias{domainData,Anno-method}
\alias{domainNames}
\alias{domainNames,Anno-method}
\alias{matrix2Anno}
\alias{show,Anno-method}
\alias{tData}
\alias{tData,Anno-method}
\alias{termData}
\alias{termData,Anno-method}
\alias{termNames}
\alias{termNames,Anno-method}
\title{Methods defined for class Anno}
\usage{
\S4method{dim}{Anno}(x)

\S4method{annoData}{Anno}(x)

\S4method{termData}{Anno}(x)

\S4method{domainData}{Anno}(x)

\S4method{tData}{Anno}(object)

\S4method{dData}{Anno}(object)

\S4method{termNames}{Anno}(object)

\S4method{domainNames}{Anno}(object)

\S4method{show}{Anno}(object)
}
\arguments{
\item{x}{an object of class \code{Anno}}

\item{object}{an object of class \code{Anno}}
}
\description{
Methods defined for class \code{Anno}.
}
\seealso{
\code{\link{Anno-class}}
}
\keyword{methods}

