% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ClassMethod-dcGOR.r
\docType{class}
\name{AnnoData-class}
\alias{AnnoData}
\alias{AnnoData-class}
\title{Definition for VIRTUAL S4 class AnnoData}
\value{
Class AnnoData
}
\description{
\code{AnnoData} is union of other classes: either matrix or dgCMatrix
(a sparse matrix in the package Matrix). It is used as a virtual class
}
\seealso{
\code{\link{Anno-class}}
}
\keyword{S4}
\keyword{classes}

