% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dcFunArgs}
\alias{dcFunArgs}
\title{Function to assign (and evaluate) arguments with default values for an input function}
\usage{
dcFunArgs(fun, verbose = T)
}
\arguments{
\item{fun}{an input function name (character string)}

\item{verbose}{logical to indicate whether the messages will be
displayed in the screen. By default, it sets to TRUE for display}
}
\value{
a list containing arguments and their default values
}
\description{
\code{dcFunArgs} is supposed to assign (and evaluate) arguments with
default values for an input function.
}
\note{
This function is potentially useful when debugging. Because the
developer does not have to specify default values for all arguments
except those arguments are of interest
}
\examples{
fun <- "dcAlgoPredictMain"
dcFunArgs(fun)
}
\seealso{
\code{\link{dcAlgoPredictMain}}
}

