\docType{methods}
\name{InfoDataFrame-method}
\alias{Data}
\alias{Data,InfoDataFrame-method}
\alias{InfoDataFrame-method}
\alias{colNames}
\alias{colNames,InfoDataFrame-method}
\alias{data.frame2InfoDataFrame}
\alias{dim,InfoDataFrame-method}
\alias{dimLabels}
\alias{dimLabels,InfoDataFrame-method}
\alias{ncol}
\alias{ncol,InfoDataFrame-method}
\alias{nrow}
\alias{nrow,InfoDataFrame-method}
\alias{rowNames}
\alias{rowNames,InfoDataFrame-method}
\alias{show,InfoDataFrame-method}
\title{Methods defined for class InfoDataFrame}
\usage{
dimLabels(x)

\S4method{dimLabels}{InfoDataFrame}(x)

\S4method{dim}{InfoDataFrame}(x)

\S4method{nrow}{InfoDataFrame}(x)

\S4method{ncol}{InfoDataFrame}(x)

\S4method{rowNames}{InfoDataFrame}(x)

\S4method{colNames}{InfoDataFrame}(x)

\S4method{Data}{InfoDataFrame}(x)

\S4method{show}{InfoDataFrame}(object)
}
\arguments{
\item{x}{an object of class \code{InfoDataFrame}}

\item{object}{an object of class \code{InfoDataFrame}}
}
\description{
Methods defined for class \code{InfoDataFrame}.
}
\seealso{
\code{\link{InfoDataFrame-class}}
}
\keyword{methods}

