% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{Anno-method}
\alias{Anno-method}
\alias{[,Anno,ANY,ANY,ANY-method}
\alias{[,Anno-method}
\alias{annoData}
\alias{annoData,Anno-method}
\alias{dData}
\alias{dData,Anno-method}
\alias{dgCMatrix2Anno}
\alias{dim,Anno-method}
\alias{domainData}
\alias{domainData,Anno-method}
\alias{domainNames}
\alias{domainNames,Anno-method}
\alias{matrix2Anno}
\alias{show,Anno-method}
\alias{tData}
\alias{tData,Anno-method}
\alias{termData}
\alias{termData,Anno-method}
\alias{termNames}
\alias{termNames,Anno-method}
\title{Methods defined for S4 class Anno}
\usage{
\S4method{dim}{Anno}(x)

\S4method{annoData}{Anno}(x)

\S4method{termData}{Anno}(x)

\S4method{domainData}{Anno}(x)

\S4method{tData}{Anno}(object)

\S4method{dData}{Anno}(object)

\S4method{termNames}{Anno}(object)

\S4method{domainNames}{Anno}(object)

\S4method{show}{Anno}(object)

\S4method{[}{Anno,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{an object of class \code{Anno}}

\item{object}{an object of class \code{Anno}}

\item{i}{an index}

\item{j}{an index}

\item{...}{additional parameters}

\item{drop}{For matrices and arrays.  If \code{TRUE} the result is
coerced to the lowest possible dimension (see the examples).  This
only works for extracting elements, not for the replacement.  See
\code{\link{drop}} for further details.
}
}
\description{
Methods defined for class \code{Anno}.
}
\seealso{
\code{\link{Anno-class}}
}
\keyword{S4}
\keyword{methods}

