% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ClassMethod-dcGOR.r
\docType{methods}
\name{InfoDataFrame-method}
\alias{Data}
\alias{Data,InfoDataFrame-method}
\alias{InfoDataFrame-method}
\alias{[,InfoDataFrame,ANY,ANY,ANY-method}
\alias{[,InfoDataFrame-method}
\alias{colNames}
\alias{colNames,InfoDataFrame-method}
\alias{data.frame2InfoDataFrame}
\alias{dim,InfoDataFrame-method}
\alias{dimLabels}
\alias{dimLabels,InfoDataFrame-method}
\alias{ncol}
\alias{ncol,InfoDataFrame-method}
\alias{nrow}
\alias{nrow,InfoDataFrame-method}
\alias{rowNames}
\alias{rowNames,InfoDataFrame-method}
\alias{show,InfoDataFrame-method}
\title{Methods defined for S4 class InfoDataFrame}
\usage{
dimLabels(x)

\S4method{dimLabels}{InfoDataFrame}(x)

\S4method{dim}{InfoDataFrame}(x)

\S4method{nrow}{InfoDataFrame}(x)

\S4method{ncol}{InfoDataFrame}(x)

\S4method{rowNames}{InfoDataFrame}(x)

\S4method{colNames}{InfoDataFrame}(x)

\S4method{Data}{InfoDataFrame}(x)

\S4method{show}{InfoDataFrame}(object)

\S4method{[}{InfoDataFrame,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{an object of class \code{InfoDataFrame}}

\item{object}{an object of class \code{InfoDataFrame}}

\item{i}{an index}

\item{j}{an index}

\item{...}{additional parameters}

\item{drop}{a logic for matrices and arrays. If TRUE the result is
coerced to the lowest possible dimension. This only works for
extracting elements, not for the replacement}
}
\description{
Methods defined for class \code{InfoDataFrame}.
}
\seealso{
\code{\link{InfoDataFrame-class}}
}
\keyword{S4}
\keyword{methods}

